# docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/block/{blockID}/tx/{txID}": {
            "get": {
                "description": "Get tx by hash or index number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "GetTx handler",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Block number, or 'latest'",
                        "name": "blockID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tx index, or tx hash",
                        "name": "txID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TxResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.TxResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.TxResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.ErrorForm": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "api.TxResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.Transaction"
                },
                "error": {
                    "$ref": "#/definitions/api.ErrorForm"
                },
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "domain.Transaction": {
            "type": "object",
            "properties": {
                "blockHash": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "gas": {
                    "type": "string"
                },
                "gasPrice": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "transactionIndex": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}